Reporters: extract SourceManager to a separate class
