+++ lib/CMakeLists.txt
@@ -54,6 +54,7 @@ set(mull_sources
 
   JunkDetection/CXX/CXXJunkDetector.cpp
 
+  Reporters/SourceManager.cpp
   Reporters/SQLiteReporter.cpp
   Reporters/TimeReporter.cpp
   SourceLocation.cpp
